seta-ui/seta-flask-server/requirements.txt
flask==2.1.3
flask-cors==3.0.10
Pillow==9.1.0
pymongo==4.0.2
python-cas==1.6.0
pyjwt==2.1.0
gunicorn==20.1.0
flask_jwt_extended==4.3.1
setproctitle==1.2.2
pycryptodome==3.14.1
flask_apscheduler==1.12.3
python-secrets==22.6.1
flask-pymongo==2.3.0
log4mongo==1.7.0
flask-restx==1.0.3
python-interface==1.6.1
injector==0.20.1
flask-injector==0.14.0